A lot of changes
Added a simple Test program file to the Compiler crate, which has been taken from a small Assignment and serves as a small sample size to
test the Compiler against an actual Program that works and was written without the goal of testing a compiler.

Updated the Tokendata a bit

Added a lot more logic to the Syntax-Parser, which can now handle a lot more of the common parts you will use in a C-Program.

Updated the Preprocessor as well to handle Conditional Compilation and in general be more robust and possibly faster and more efficient as it
now returns an Iterator over the Tokens, which should reduce memory consumption as well as cpu usage as it now only evaluates the Tokens as
they are needed and they are then immediately consumed