Added first Semantics stuff
Added the initial stuff for the Semenatics part as the next part of the Compiler.
Then also improved the Error messages of the Compiler Binary to be actually more helpful than they were previously.
Also readjusted some more exports for the syntax and general stuff for easier usage in the rest of the compiler-arch