Better errors and better coverage
Improved the handling of error messages in the Binary Crate which is now way more ergonomic and makes it easier to
write good error messages based on the Data that gets given to it by the actual compiler.

Then overall improved the Way that Data and Spans are handled in the Compiler and now preserves more of it to hopefully
generate better error messages later on, when needed. This however is not done by any means and definetly needs more work
as there are certain areas that just panic for now because I currently dont have the needed Data in that place to return
the proper Errors, which is not what we want