Worked more on the IR
Worked some more on the IR to improve its usability and also added a way for it to be easily converted into a dot graphviz file
to visualize the structure, for debugging.

The Conversion from AAST to IR now works a bit better and should support if-else-statements as intended, although the test cases
for the else parts are not entirely done yet. The Next part would be adding support for basic loops