Fixed a bunch more stuff and moved tests around
Fixed the Debug output for the IR-Blocks and Program to not hang when printing cycles, although this can still be improved to make
it a bit prettier, but at least it works now and is actually pretty useable.

Moves the AAST to IR conversion tests around to better seperate them from the Rest of the Tests in the Semantic crate. Then also added
the first test to deal with while loops, which should work from the logic side but the test itself is not entirely finished