Worked on the first iteration of Backends
This includes a first initial Backend for M1 Macs.

Alongside the Backend, Im also building up a set of utilities that can be used by all the Backends to make it easier
to add a new one. These only really include a Register Allocator for but can be expanded in the Future