A lot more fixes and additions
Finally fixed loops, I think, and also now changed for loops so that they are first converted to while loops and then these while loops
are converted to IR as this should simplify the IR generation as well as reducing the duplication of generation for loops.

Then also worked a lot more on the Backend to now support, local variables properly, arrays and structs should now also work, and function
calls are now working on a very basic level (without arguments that is).