Backend for Calculator now Works + Better CLI
Fixed up some more stuff for the Calculator Backend to get it working as expected and it can now actually compile basic Programs for
it, which is good to know.

Then also added Clap to improve the Compilers CLI a bit more and provide ways for the User to easily configure things like that Target
or the extra Paths to search for includes.