A lot of improvements
Worked a lot on the Backend of the Compiler to make it more Robust and add more features.
Currently the focus is on the Register allocator and spilling of Variables to get them to work properly
and produce pretty good Code, not quite "perfect" but really solid code. This currently requires a lot of
restructuring of the Code and changing how all of this is handled to fit all the needed Parts into it without
making it to convoluted and messy.
There also needs to be more documentation for this spilling logic as it can be quite confusing, however once it
is done, the generated IR/Code should be pretty good and handle the different Contexts (in a loop, conditional, etc.)
with a good strategy for each case