More work on Register Allocation
Switched the spilling algorithm to one based on a Paper from the Uni Kalsruhe, which seems to be working very well and
is way more promising than whatever I was building up before.
The implementation is not really fleshed out in any way, shape or form, but is enough for simple sections as a sort of
POC.
However because this algorithm performs a lot better, we need to improve the Tests that are used to determine if spilling
happens correctly.