Updated Interference Graphs and Jumps
Updated Jumps to also store some Metadata, which for now just indicates what kind of Jump it is, linear, loop, loopbreak,
and should not yet be relied on. However this should help the Register Allocator/Spilling as it gives a bit more information
easily to inform about which variables to prioritize when encountering loops.

Then also updated Interference Graphs so now Operands don't interfere with the Result, if its the Operands last use, which should
be more correct for load/store architectures, which are the Primary targest of this Compiler