Added Semantics Fuzzer
Added a Fuzzer to the Semantics crate as well and then used it to discover some more obvious crashes in my Semantics
Crate. However this is nowhere near good as it still first needs to parse the syntax and then give the result to the
semantics crate which both takes longer and also means we could find bugs in the syntax crate which is not the goal.

Then also fixed some more crashes found in the syntax and semantic crate